fix the bug of wrong result when comparing two doubles

+++ test.c
@@ -19,7 +19,7 @@ static int test_pass = 0;
     } while(0)
 
 #define EXPECT_EQ_INT(expect, actual) EXPECT_EQ_BASE((expect) == (actual), expect, actual, "%d")
-#define EXPECT_EQ_DOUBLE(expect, actual) EXPECT_EQ_BASE((expect) == (actual), expect, actual, "%.17g")
+#define EXPECT_EQ_DOUBLE(expect, actual) EXPECT_EQ_BASE((double)(expect) == (actual), expect, actual, "%.17g")
 
 #define TEST_LITERAL(expected_type, json)\
     do{\
